Move types and hooks from App component
