+++ src/App.tsx
@@ -1,26 +1,12 @@
-import React, { useState } from 'react';
+import React from 'react';
 import UserLogin from './components/UserLogin';
 import ChatRoom from './components/ChatRoom';
 import Header from './components/Header';
-
-export type SetIsLoggedIn = React.Dispatch<
-  React.SetStateAction<loggedInDataType>
->;
-
-export interface loggedInDataType {
-  logUsername: string;
-  isLoggedIn: boolean;
-}
-
-type AppType = () => JSX.Element;
+import useLogin from './hooks/useLogin';
+import { AppType } from './types';
 
 const App: AppType = () => {
-  const logUsername = sessionStorage.getItem('logUsername');
-
-  const [loggedInData, setLoggedInData] = useState<loggedInDataType>({
-    logUsername: logUsername || '',
-    isLoggedIn: !!logUsername || false,
-  });
+  const { loggedInData, setLoggedInData } = useLogin();
 
   return (
     <>
